+++ src/components/mote/container/MotelandingssideContainer.js
@@ -8,10 +8,7 @@ import AppSpinner from "../../AppSpinner";
 import Feilmelding from "../../Feilmelding";
 import Motelandingsside from "../components/Motelandingsside";
 import { hentBegrunnelseTekst } from "../../../utils/tilgangUtils";
-import {
-  isUnfinishedMotebehovTask,
-  isUnfinishedMoterTask,
-} from "../../../utils/GlobalNavigasjonUtils";
+import { isUnfinishedMoterTask } from "../../../utils/GlobalNavigasjonUtils";
 
 const texts = {
   pageTitle: "Møtelandingsside",
@@ -30,7 +27,6 @@ export class MotelandingssideSide extends Component {
       henter,
       hentingFeilet,
       mote,
-      skalViseMotebehovPrikk,
       skalViseMotePrikk,
       tilgang,
     } = this.props;
@@ -59,7 +55,6 @@ export class MotelandingssideSide extends Component {
             <Motelandingsside
               fnr={fnr}
               mote={mote}
-              skalViseMotebehovPrikk={skalViseMotebehovPrikk}
               skalViseMotePrikk={skalViseMotePrikk}
             />
           );
@@ -75,7 +70,6 @@ MotelandingssideSide.propTypes = {
   hentingFeilet: PropTypes.bool,
   hentMoter: PropTypes.func,
   mote: PropTypes.object,
-  skalViseMotebehovPrikk: PropTypes.bool,
   skalViseMotePrikk: PropTypes.bool,
   tilgang: PropTypes.object,
 };
@@ -85,7 +79,6 @@ export const mapStateToProps = (state, ownProps) => {
     return mote.status !== "AVBRUTT";
   })[0];
 
-  const skalViseMotebehovPrikk = isUnfinishedMotebehovTask(state.motebehov);
   const skalViseMotePrikk = isUnfinishedMoterTask(state.moter);
 
   return {
@@ -93,7 +86,6 @@ export const mapStateToProps = (state, ownProps) => {
     henter: state.tilgang.henter || state.moter.henter,
     hentingFeilet: state.tilgang.hentingFeilet,
     mote: aktivtMote,
-    skalViseMotebehovPrikk,
     skalViseMotePrikk,
     tilgang: state.tilgang.data,
   };
