+++ src/styles/_motebehovKvittering.less
@@ -1,7 +1,5 @@
 .motebehovKvitteringInnhold {
-  padding-top: 0.5em;
-  margin: 1em 0;
-  padding-bottom: 0.5em;
+  padding-bottom: 1em;
   label {
     margin-left: 0.75em;
     display: block;
