+++ src/styles/_motelandingsside.less
@@ -8,7 +8,7 @@
   border: 0;
   background-color: @white;
   padding: 1.5em 3em 1.5em 2em;
-  margin-bottom: 0.2em;
+  margin-bottom: 2em;
   color: @navMorkGra;
   border-left: 0;
   background-image: url("/sykefravaer/img/svg/pilVidere--graa.svg");
@@ -17,12 +17,6 @@
   background-size: 1rem 1rem;
   transition: background-position ease-in-out 0.1s;
   transform: translate3d(0, 0, 0);
-}
-
-.motelandingssidepanel {
-  display: flex;
-  margin-left: -2em;
-  list-style-type: none;
 }
 
 .motelandingssidepanel__tekst {
